Use geos PreparedGeometry for faster intersections
Without PreparedGeometry, every Geometry::intersects(other_geom) call
has to do a bunch of work to set up for the intersection. a
PreparedGeometry is able to do much of that work once, and reuse the
resulting data structures during each intersection test. It's a _lot_
faster.

See http://blog.cleverelephant.ca/2008/10/postgis-performance-prepared-geometry.html
for an example of the performance boost.